Add cluster_dims param to MatmulParams
Following #3557 we can specify the cluster size for our fusions.
Currently we don't do anything explicitly with CGAs, but this can help
guarantee that tiles are scheduled onto GPCs in pairs. Each GPC has a
number of TPCs, each of which holds 2 SMs, so this lets us take
advantage of caching at the TPC and GPC level for operand loads, in
addition to L2.

+++ csrc/codegen.cpp
@@ -276,8 +276,7 @@ class CudaKernelGenerator : private kir::ConstIrVisitor {
     code_ << "__global__ void ";
     if (kernel_->hasManaged("cluster_dims")) {
       auto cluster_dims =
-          kernel_->getManaged<std::tuple<int64_t, int64_t, int64_t>>(
-              "cluster_dims");
+          kernel_->getManaged<std::tuple<int, int, int>>("cluster_dims");
       code_ << "__cluster_dims__(" << std::get<0>(cluster_dims) << ", "
             << std::get<1>(cluster_dims) << ", " << std::get<2>(cluster_dims)
             << ") ";
